[APPC-1634] Completed initial payment methods portrait layout. Added radio button pngs
